requirements.txt
Django
djangorestframework
django-filter
python-decouple
cloudinary
django-cloudinary-storage
celery
redis
